Improved Coral
- Took out coral glow script
- Reset shader on coral glow prefab
+ Coral Glow New script implemented
+ Changed logic on how Coral Glow New's health is implemented
