+++ Assets/Scripts/Corals/CoralGlowNew.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 12546c0b58257c348ad88c48b7bb500c
+guid: 196ca3a1f24b8b046bd470bff3a18aea
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
